images: refactored code into common utility module
This means that we don't have to import utility functions from other commands.
